Combine expansion filters into single Python module
