+++ openmc/__init__.py
@@ -15,10 +15,7 @@ from openmc.surface import *
 from openmc.universe import *
 from openmc.lattice import *
 from openmc.filter import *
-from openmc.filter_legendre import *
-from openmc.filter_spatial_legendre import *
-from openmc.filter_spherical_harmonics import *
-from openmc.filter_zernike import *
+from openmc.filter_expansion import *
 from openmc.trigger import *
 from openmc.tally_derivative import *
 from openmc.tallies import *
