--- openmc/filter_legendre.py
@@ -1,90 +1,0 @@
-from numbers import Integral
-from xml.etree import ElementTree as ET
-
-import numpy as np
-import pandas as pd
-
-import openmc.checkvalue as cv
-from . import Filter
-
-
-class LegendreFilter(Filter):
-    r"""Score Legendre expansion moments up to specified order.
-
-    This filter allows scores to be multiplied by Legendre polynomials of the
-    change in particle angle ($\mu$) up to a user-specified order.
-
-    Parameters
-    ----------
-    order : int
-        Maximum Legendre polynomial order
-    filter_id : int or None
-        Unique identifier for the filter
-
-    Attributes
-    ----------
-    order : int
-        Maximum Legendre polynomial order
-    id : int
-        Unique identifier for the filter
-    num_bins : int
-        The number of filter bins
-
-    """
-
-    def __init__(self, order, filter_id=None):
-        self.order = order
-        self.bins = ['P{}'.format(i) for i in range(order + 1)]
-        self.id = filter_id
-
-    def __hash__(self):
-        string = type(self).__name__ + '\n'
-        string += '{: <16}=\t{}\n'.format('\tOrder', self.order)
-        return hash(string)
-
-    def __repr__(self):
-        string = type(self).__name__ + '\n'
-        string += '{: <16}=\t{}\n'.format('\tOrder', self.order)
-        string += '{: <16}=\t{}\n'.format('\tID', self.id)
-        return string
-
-    @property
-    def order(self):
-        return self._order
-
-    @order.setter
-    def order(self, order):
-        cv.check_type('Legendre order', order, Integral)
-        cv.check_greater_than('Legendre order', order, 0, equality=True)
-        self._order = order
-
-    @classmethod
-    def from_hdf5(cls, group, **kwargs):
-        if group['type'].value.decode() != cls.short_name.lower():
-            raise ValueError("Expected HDF5 data for filter type '"
-                             + cls.short_name.lower() + "' but got '"
-                             + group['type'].value.decode() + " instead")
-
-        filter_id = int(group.name.split('/')[-1].lstrip('filter '))
-
-        out = cls(group['order'].value, filter_id)
-
-        return out
-
-    def to_xml_element(self):
-        """Return XML Element representing the filter.
-
-        Returns
-        -------
-        element : xml.etree.ElementTree.Element
-            XML element containing Legendre filter data
-
-        """
-        element = ET.Element('filter')
-        element.set('id', str(self.id))
-        element.set('type', self.short_name.lower())
-
-        subelement = ET.SubElement(element, 'order')
-        subelement.text = str(self.order)
-
-        return element
